+++ common/src/main/java/dev/ftb/mods/ftbessentials/util/Leaderboard.java
@@ -16,6 +16,8 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Leaderboard<N extends Number> {
 	private static final DecimalFormat DECIMAL_FORMAT = Util.make(new DecimalFormat("########0.00"), decimalFormat -> decimalFormat.setDecimalFormatSymbols(DecimalFormatSymbols.getInstance(Locale.ROOT)));
@@ -102,6 +104,17 @@ public class Leaderboard<N extends Number> {
 
 	public String getName() {
 		return name;
+	}
+
+	/**
+	 * Remove the underscores from the name and capitalize the first letter of each word
+	 * @return the formatted name
+	 */
+	public String formattedName() {
+		var parts = this.name.split("_");
+		return Stream.of(parts)
+				.map(s -> Character.toTitleCase(s.charAt(0)) + s.substring(1).toLowerCase())
+				.collect(Collectors.joining(" "));
 	}
 
 	public Leaderboard<N> withValueGetter(Function<ServerStatsCounter, N> v) {